+++ packages/components/src/Banner/Banner.css.d.ts
@@ -1,2 +1,6 @@
 export const flash: string;
+export const medium: string;
+export const bannerContent: string;
+export const dismissibleSpacing: string;
+export const bannerAction: string;
 export const closeButton: string;