+++ packages/components/src/FormField/FormField.test.tsx
@@ -1,36 +1,37 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import { cleanup, fireEvent, render } from "@testing-library/react";
+import { cleanup, fireEvent, render, waitFor } from "@testing-library/react";
 import { FormField } from ".";
 
 afterEach(cleanup);
 it("renders correctly with no props", () => {
   const tree = renderer.create(<FormField />).toJSON();
   expect(tree).toMatchInlineSnapshot(`
-        <div
-          className="wrapper"
-          style={
-            Object {
-              "--formField-maxLength": undefined,
-            }
-          }
-        >
-          <label
-            className="label"
-            htmlFor="123e4567-e89b-12d3-a456-426655440001"
-          >
-             
-          </label>
-          <input
-            className="formField"
-            id="123e4567-e89b-12d3-a456-426655440001"
-            onBlur={[Function]}
-            onChange={[Function]}
-            onFocus={[Function]}
-            type="text"
-          />
-        </div>
-    `);
+    <div
+      className="wrapper"
+      style={
+        Object {
+          "--formField-maxLength": undefined,
+        }
+      }
+    >
+      <label
+        className="label"
+        htmlFor="123e4567-e89b-12d3-a456-426655440001"
+      >
+         
+      </label>
+      <input
+        className="formField"
+        id="123e4567-e89b-12d3-a456-426655440001"
+        onBlur={[Function]}
+        onChange={[Function]}
+        onFocus={[Function]}
+        onKeyDown={[Function]}
+        type="text"
+      />
+    </div>
+  `);
 });
 
 it("renders correctly with a placeholder", () => {
@@ -58,6 +59,7 @@ it("renders correctly with a placeholder", () => {
         onBlur={[Function]}
         onChange={[Function]}
         onFocus={[Function]}
+        onKeyDown={[Function]}
         type="text"
       />
     </div>
@@ -87,6 +89,7 @@ it("renders correctly as small", () => {
         onBlur={[Function]}
         onChange={[Function]}
         onFocus={[Function]}
+        onKeyDown={[Function]}
         type="text"
       />
     </div>
@@ -116,6 +119,7 @@ it("renders correctly in a readonly state", () => {
         onBlur={[Function]}
         onChange={[Function]}
         onFocus={[Function]}
+        onKeyDown={[Function]}
         readOnly={true}
         type="text"
       />
@@ -147,6 +151,7 @@ it("renders correctly in a disabled state", () => {
         onBlur={[Function]}
         onChange={[Function]}
         onFocus={[Function]}
+        onKeyDown={[Function]}
         type="text"
       />
     </div>
@@ -154,43 +159,33 @@ it("renders correctly in a disabled state", () => {
 });
 
 it("renders a field with error", () => {
-  const tree = renderer
-    .create(
-      <FormField value="wrong!" errorMessage="Enter a value that is correct" />,
-    )
-    .toJSON();
+  const tree = renderer.create(<FormField value="wrong!" />).toJSON();
   expect(tree).toMatchInlineSnapshot(`
-    Array [
-      <p
-        className="base regular base red"
-      >
-        Enter a value that is correct
-      </p>,
-      <div
-        className="wrapper invalid"
-        style={
-          Object {
-            "--formField-maxLength": undefined,
-          }
+    <div
+      className="wrapper"
+      style={
+        Object {
+          "--formField-maxLength": undefined,
         }
+      }
+    >
+      <label
+        className="label"
+        htmlFor="123e4567-e89b-12d3-a456-426655440006"
       >
-        <label
-          className="label"
-          htmlFor="123e4567-e89b-12d3-a456-426655440006"
-        >
-           
-        </label>
-        <input
-          className="formField"
-          id="123e4567-e89b-12d3-a456-426655440006"
-          onBlur={[Function]}
-          onChange={[Function]}
-          onFocus={[Function]}
-          type="text"
-          value="wrong!"
-        />
-      </div>,
-    ]
+         
+      </label>
+      <input
+        className="formField"
+        id="123e4567-e89b-12d3-a456-426655440006"
+        onBlur={[Function]}
+        onChange={[Function]}
+        onFocus={[Function]}
+        onKeyDown={[Function]}
+        type="text"
+        value="wrong!"
+      />
+    </div>
   `);
 });
 
@@ -234,25 +229,139 @@ test("it should call the validation handler when typing a new value", () => {
   render(
     <FormField
       name="Got milk?"
-      onValidate={validationHandler}
+      onValidation={validationHandler}
       placeholder="I hold places."
     />,
   );
 
-  expect(validationHandler).toHaveBeenCalledWith("pass", "");
+  expect(validationHandler).toHaveBeenCalled();
+  expect(validationHandler).toHaveBeenCalledWith(undefined);
 });
 
-test("it should call the validation handler with a fail status when there's an error", () => {
+test("it should call the validation handler with a message when there is an error", async () => {
   const validationHandler = jest.fn();
+  const validate = val => (val !== "Bob" ? "message" : "");
 
-  render(
+  const { getByLabelText } = render(
     <FormField
+      type="text"
       name="Got milk?"
-      onValidate={validationHandler}
+      onValidation={validationHandler}
       placeholder="I hold places"
-      errorMessage="Nope!"
+      value="test"
+      validations={{
+        validate,
+      }}
     />,
   );
 
-  expect(validationHandler).toHaveBeenCalledWith("fail", "Nope!");
+  const input = getByLabelText("I hold places");
+  input.focus();
+  fireEvent.change(input, { target: { value: "Bob" } });
+  input.blur();
+
+  expect(validationHandler).toHaveBeenCalled();
+  await waitFor(() => {
+    expect(validationHandler).toHaveBeenCalledWith("message");
+  });
+});
+
+test("it should handle when the enter key is pressed", () => {
+  const enterHandler = jest.fn();
+  const placeholder = "Milk heals bones";
+
+  const { getByLabelText } = render(
+    <FormField
+      name="Enter the milk house"
+      onEnter={enterHandler}
+      placeholder={placeholder}
+    />,
+  );
+
+  fireEvent.keyDown(getByLabelText(placeholder), {
+    key: "Enter",
+    code: "Enter",
+  });
+
+  expect(enterHandler).toHaveBeenCalledTimes(1);
+
+  fireEvent.keyDown(getByLabelText(placeholder), {
+    key: "Enter",
+    code: "Enter",
+  });
+
+  expect(enterHandler).toHaveBeenCalledTimes(2);
+});
+
+test("it should not handle when the shift key and enter key are pressed", () => {
+  const enterHandler = jest.fn();
+  const placeholder = "Milk heals bones";
+
+  const { getByLabelText } = render(
+    <FormField
+      name="Enter the milk house"
+      onEnter={enterHandler}
+      placeholder={placeholder}
+    />,
+  );
+
+  fireEvent.keyDown(getByLabelText(placeholder), {
+    key: "Enter",
+    code: "Enter",
+    shiftKey: true,
+  });
+
+  expect(enterHandler).toHaveBeenCalledTimes(0);
+});
+
+test("it should not handle when the shift key and control key are pressed", () => {
+  const enterHandler = jest.fn();
+  const placeholder = "Milk heals bones";
+
+  const { getByLabelText } = render(
+    <FormField
+      name="Enter the milk house"
+      onEnter={enterHandler}
+      placeholder={placeholder}
+    />,
+  );
+
+  fireEvent.keyDown(getByLabelText(placeholder), {
+    key: "Enter",
+    code: "Enter",
+    ctrlKey: true,
+  });
+
+  expect(enterHandler).toHaveBeenCalledTimes(0);
+});
+
+test("it should not have a name by default", () => {
+  const { getByLabelText } = render(<FormField placeholder="foo" />);
+  expect(getByLabelText("foo")).not.toHaveAttribute("name");
+});
+
+test("it should use the name prop when set", () => {
+  const { getByLabelText } = render(
+    <FormField placeholder="foo" name="dillan" />,
+  );
+  expect(getByLabelText("foo")).toHaveAttribute("name", "dillan");
+});
+
+test("it should generate a name if validations are set", () => {
+  const { getByLabelText } = render(
+    <FormField placeholder="foo" validations={{ required: true }} />,
+  );
+  const input = getByLabelText("foo");
+  const name = input.getAttribute("name");
+  expect(name).toContain("generatedName--");
+});
+
+test("it should set the inputMode when the keyboard prop is set", () => {
+  const keyboardMode = "numeric";
+  const { getByLabelText } = render(
+    <FormField placeholder="foo" keyboard={keyboardMode} />,
+  );
+  const input = getByLabelText("foo");
+  const name = input.getAttribute("inputMode");
+  expect(name).toContain(keyboardMode);
 });