+++ packages/components/src/FormatFile/FormatFile.css.d.ts
@@ -1,0 +1,6 @@
+export const formatFile: string;
+export const imageBlock: string;
+export const icon: string;
+export const progress: string;
+export const contentBlock: string;
+export const actionBlock: string;