+++ packages/components/src/InputFile/InputFile.css.d.ts
@@ -1,0 +1,3 @@
+export const dropZoneBase: string;
+export const dropZone: string;
+export const active: string;