+++ packages/components/src/InputText/InputText.tsx
@@ -2,6 +2,11 @@ import React, { Ref, createRef, forwardRef, useImperativeHandle } from "react";
 import { XOR } from "ts-xor";
 import { FormField, FormFieldProps } from "../FormField";
 
+interface RowRange {
+  min: number;
+  max: number;
+}
+
 /**
  * The following is the same as:
  *   type BaseProps = Omit<FormFieldProps, "type" | "children">;
@@ -16,6 +21,7 @@ type BaseProps = Pick<
 export interface InputTextRef {
   insert(text: string): void;
   blur(): void;
+  focus(): void;
 }
 
 interface MultilineProps extends BaseProps {
@@ -25,9 +31,12 @@ interface MultilineProps extends BaseProps {
   readonly multiline: true;
 
   /**
-   * Specifies the visible height of a long answer form field.
+   * Specifies the visible height of a long answer form field. Can be in the
+   * form of a single number to set a static height, or an object with a min
+   * and max keys indicating the minimum number of visible rows, and the
+   * maximum number of visible rows.
    */
-  readonly rows?: number;
+  readonly rows?: number | RowRange;
 }
 
 type InputTextPropOptions = XOR<BaseProps, MultilineProps>;
@@ -37,6 +46,7 @@ function InputTextInternal(
   ref: Ref<InputTextRef>,
 ) {
   const inputRef = createRef<HTMLTextAreaElement | HTMLInputElement>();
+  const rowRange = getRowRange();
 
   useImperativeHandle(ref, () => ({
     insert: (text: string) => {
@@ -48,15 +58,72 @@ function InputTextInternal(
         input.blur();
       }
     },
+    focus: () => {
+      const input = inputRef.current;
+      if (input) {
+        input.focus();
+      }
+    },
   }));
 
   return (
     <FormField
-      type={props.multiline ? "textarea" : "text"}
-      ref={inputRef}
       {...props}
+      type={props.multiline ? "textarea" : "text"}
+      inputRef={inputRef}
+      onChange={handleChange}
+      rows={rowRange.min}
     />
   );
+
+  function handleChange(newValue: string) {
+    props.onChange && props.onChange(newValue);
+
+    if (inputRef && inputRef.current instanceof HTMLTextAreaElement) {
+      resize(inputRef.current);
+    }
+  }
+
+  function getRowRange(): RowRange {
+    if (props.rows === undefined) {
+      return { min: 3, max: 3 };
+    } else if (typeof props.rows === "object") {
+      return { min: props.rows.min, max: props.rows.max };
+    } else {
+      return { min: props.rows, max: props.rows };
+    }
+  }
+
+  function resize(textArea: HTMLTextAreaElement) {
+    if (rowRange.min === rowRange.max) return;
+
+    textArea.style.height = "auto";
+    textArea.style.height = textAreaHeight(textArea) + "px";
+  }
+
+  function textAreaHeight(textArea: HTMLTextAreaElement) {
+    const {
+      lineHeight,
+      borderBottomWidth,
+      borderTopWidth,
+      paddingBottom,
+      paddingTop,
+    } = window.getComputedStyle(textArea);
+
+    const maxHeight =
+      rowRange.max * parseFloat(lineHeight) +
+      parseFloat(borderTopWidth) +
+      parseFloat(borderBottomWidth) +
+      parseFloat(paddingTop) +
+      parseFloat(paddingBottom);
+
+    const scrollHeight =
+      textArea.scrollHeight +
+      parseFloat(borderTopWidth) +
+      parseFloat(borderBottomWidth);
+
+    return Math.min(scrollHeight, maxHeight);
+  }
 
   function insertText(text: string) {
     const input = inputRef.current;