+++ packages/components/src/RadioGroup/RadioGroup.css.d.ts
@@ -1,3 +1,4 @@
+export const radioGroup: string;
 export const input: string;
 export const label: string;
-export const radioGroup: string;
+export const description: string;