+++ packages/components/src/Toast/Toast.css.d.ts
@@ -1,0 +1,7 @@
+export const wrapper: string;
+export const container: string;
+export const toast: string;
+export const slice: string;
+export const icon: string;
+export const button: string;
+export const action: string;