+++ packages/design/CHANGELOG.md
@@ -8,6 +8,90 @@ menu: Changelog
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.3](https://github.com/GetJobber/atlantis/compare/@jobber/design@0.9.2...@jobber/design@0.9.3) (2021-01-11)
+
+**Note:** Version bump only for package @jobber/design
+
+
+
+
+
+## [0.9.2](https://github.com/GetJobber/atlantis/compare/@jobber/design@0.9.1...@jobber/design@0.9.2) (2020-12-03)
+
+
+### Bug Fixes
+
+* **components:** Change `Table` borders from box-shadow to pseudo element ([#427](https://github.com/GetJobber/atlantis/issues/427)) ([7648514](https://github.com/GetJobber/atlantis/commit/76485149e583a0711f3592881035df550aef3393))
+
+
+
+
+
+## [0.9.1](https://github.com/GetJobber/atlantis/compare/@jobber/design@0.9.0...@jobber/design@0.9.1) (2020-10-07)
+
+**Note:** Version bump only for package @jobber/design
+
+
+
+
+
+# [0.9.0](https://github.com/GetJobber/atlantis/compare/@jobber/design@0.8.1...@jobber/design@0.9.0) (2020-10-01)
+
+
+### Features
+
+* **components:** InputAvatar & Expand variation options for InputFile ([#366](https://github.com/GetJobber/atlantis/issues/366)) ([7b07059](https://github.com/GetJobber/atlantis/commit/7b07059170487368e0c04bc0674c71c51a0f676e))
+
+
+
+
+
+## [0.8.1](https://github.com/GetJobber/atlantis/compare/@jobber/design@0.8.0...@jobber/design@0.8.1) (2020-09-17)
+
+**Note:** Version bump only for package @jobber/design
+
+
+
+
+
+# [0.8.0](https://github.com/GetJobber/atlantis/compare/@jobber/design@0.7.4...@jobber/design@0.8.0) (2020-08-11)
+
+
+### Features
+
+* **components:** Toast Component ([#326](https://github.com/GetJobber/atlantis/issues/326)) ([4e04a4d](https://github.com/GetJobber/atlantis/commit/4e04a4daa90e6773def51d66387562001806fe9c))
+
+
+
+
+
+## [0.7.4](https://github.com/GetJobber/atlantis/compare/@jobber/design@0.7.3...@jobber/design@0.7.4) (2020-08-07)
+
+**Note:** Version bump only for package @jobber/design
+
+
+
+
+
+## [0.7.3](https://github.com/GetJobber/atlantis/compare/@jobber/design@0.7.2...@jobber/design@0.7.3) (2020-07-15)
+
+
+### Bug Fixes
+
+* update icon type defs ([#312](https://github.com/GetJobber/atlantis/issues/312)) ([a571931](https://github.com/GetJobber/atlantis/commit/a571931d143616ba8ddc1b2e271bb7c734a6e307))
+
+
+
+
+
+## [0.7.2](https://github.com/GetJobber/atlantis/compare/@jobber/design@0.7.1...@jobber/design@0.7.2) (2020-07-10)
+
+**Note:** Version bump only for package @jobber/design
+
+
+
+
+
 ## [0.7.1](https://github.com/GetJobber/atlantis/compare/@jobber/design@0.7.0...@jobber/design@0.7.1) (2020-07-02)
 
 **Note:** Version bump only for package @jobber/design