+++ packages/hooks/CHANGELOG.md
@@ -8,6 +8,73 @@ menu: Changelog
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.1.0](https://github.com/GetJobber/atlantis/compare/@jobber/hooks@1.0.0...@jobber/hooks@1.1.0) (2020-12-04)
+
+
+### Features
+
+* **hooks:** Add useOnKeyDown hook ([#432](https://github.com/GetJobber/atlantis/issues/432)) ([6a30a2e](https://github.com/GetJobber/atlantis/commit/6a30a2e8ad8ea3540eaab99597b2b5298facc575))
+
+
+
+
+
+# [1.0.0](https://github.com/GetJobber/atlantis/compare/@jobber/hooks@0.3.0...@jobber/hooks@1.0.0) (2020-09-24)
+
+
+### Features
+
+* **components:** BREAKING CHANGE: FormField Validation ([#352](https://github.com/GetJobber/atlantis/issues/352)) ([a9b6221](https://github.com/GetJobber/atlantis/commit/a9b622197bf51b9f98f59280f06d3e34ed4c9636)), closes [#357](https://github.com/GetJobber/atlantis/issues/357) [#360](https://github.com/GetJobber/atlantis/issues/360)
+
+
+### BREAKING CHANGES
+
+* **components:** * add react-hook-form
+
+* Adds validation to field
+
+* Littered TODOs
+
+* Uses inputValidation component
+
+* update valid to validations prop
+
+* Makes onValidation work again
+
+* Make input validation docs work same as old atlantis
+
+* Add custom validation example
+
+* Fix the validation on input password docs
+
+* Fixed input number validation
+
+* Change FormField from forwardRef to use ref as a prop
+
+* Force a name on fields
+
+* Fix validation tests on FormField
+
+* Fix inputNumber tests
+
+* Some easy PR fixes
+
+* Styles the validation message
+
+* Add TS exceptions
+
+* Styling changes for Validation mesages
+
+* Removes Icon from validation message
+
+* Case inputRef
+
+* Cast the other ref
+
+
+
+
+
 # [0.3.0](https://github.com/GetJobber/atlantis/compare/@jobber/hooks@0.2.1...@jobber/hooks@0.3.0) (2020-07-06)